Fixed a major bug with the latest changes.
The latest implementation was not using the WriterPool functionality in Heritrix.
As a result, it would open a new connection to zookeeper each time it needed to
write a record. Furthermore, it wasn't releasing the newly creating connections
at all, so eventually Heritrix (or ZK/HBase) would lock up and nothing could
be written anymore.